+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [6.5.0] - 2026-09-07
+
+_Stable release based on [6.5.0-rc.1]._
+
 ## [6.5.0-rc.1] - 2026-09-07
 
 ### Added
@@ -877,6 +881,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[6.5.0]: https://https://github.com/internetguru/laravel-model-browser/compare/v6.4.1...v6.5.0
 [6.5.0-rc.1]: https://github.com/internetguru/laravel-model-browser/releases/tag/v6.4.1
 [6.4.1]: https://https://github.com/internetguru/laravel-model-browser/compare/v6.4.0...v6.4.1
 [6.4.0]: https://https://github.com/internetguru/laravel-model-browser/compare/v6.3.0...v6.4.0